Revert "Enable nullability in most of the remaining files in BizHawk.Common"
This reverts commit 51b62629c61c0620db4c6b9f9f3db17115c361c4.

Bizhawk does not change much; the application's features are largely the same as they were years ago, with some core work being the only notable improvements from release to release.  Yet there is always a massive amount of churn in the codebase, with multiple contributors engaing in zealous refactoring costantly.  This has to stop.  It's too much effort to police and fix regressions, for no gain -- if the code was simply left as it was, it would still work the way it did yesterday, and that would be fine.

From now on, whenever I see a commit that:
1. contains a problem
2. does not provide any real bugfix or real feature for end users
I'm just going to revert it with no further comment.  It's not worth it to sort out the good from the bad, because of #2 -- there's no value at the end of the tunnel.

If the abusive refactorings continue, I'll start reverting all refactoring commits.